backend/requirements.txt
Flask
flask-cors
requests
PyPDF2
python-dotenv
pymongo
pillow
